move v2 test class to v1 test file
